requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
async-timeout==5.0.1
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.13.3
bidict==0.23.1
billiard==4.2.1
blinker==1.9.0
celery==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coloredlogs==15.0.1
cssselect==1.3.0
distro==1.9.0
dnspython==2.7.0
eventlet==0.39.1
exceptiongroup==1.2.2
executor==23.2
Faker==37.1.0
fasteners==0.19
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
Flask-Executor==1.0.0
Flask-SocketIO==5.5.1
frozenlist==1.5.0
gevent==24.11.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
humanfriendly==10.0
idna==3.10
itsdangerous==2.2.0
jieba3k==0.35.1
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
kombu==5.5.2
lxml==5.3.2
lxml_html_clean==0.4.2
MarkupSafe==3.0.2
multidict==6.4.3
newspaper3k==0.2.8
nltk==3.9.1
openai==1.73.0
packaging==24.2
pillow==11.1.0
prompt_toolkit==3.0.50
propcache==0.3.1
property-manager==3.0
psycopg2-binary==2.9.10
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.0
python-socketio==5.13.0
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
sgmllib3k==1.0.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
tinysegmenter==0.3
tldextract==5.2.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
verboselogs==1.7
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.19.0
zope.event==5.0
zope.interface==7.2
fastapi[all] # Includes uvicorn, pydantic, etc.
uvicorn[standard]
redis
rq
rq-scheduler
psycopg2-binary # Standard PostgreSQL driver
SQLAlchemy # ORM for easier DB interaction
python-jose[cryptography] # For JWT
passlib[bcrypt] # For password hashing
python-dotenv # To manage environment variables
fastapi-mail>=1.4.1
requests>=2.25.0